// frontend/src/components/era-comparator/types.ts
export type { EraComparisonResult, EraTimelineEvent, EraKey } from '@/types'

export interface EraNameMap {
  WoodenOx: string
  Spot: string
  Cheetah: string
  ANYmal: string
}

export interface EraColorMap {
  WoodenOx: string
  Spot: string
  Cheetah: string
  ANYmal: string
}

export interface TableRow {
  key: number
  metric: string
  WoodenOx: any
  Spot: any
  Cheetah: any
  ANYmal: any
}